+++ src/main/resources/static/js/delete/deleteGymClass.js
@@ -9,6 +9,11 @@ for (const btnEliminar of btnEliminarList) {
         // Get the id of the course to delete from the data-id attribute of the button
         const id = btnEliminar.getAttribute('data-id');
 
+        // Ask the user to confirm the deletion
+        const confirmed = confirm('Are you sure you want to delete this Course?');
+        if (!confirmed) {
+            return; // Exit the function without doing anything
+        }
         // Send a DELETE request to the REST API to delete the course with the specified id
         fetch(`/api/course?courseId=${encodeURIComponent(id)}`,  {
             method: 'DELETE'
